Print the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -15,7 +15,7 @@ int main(void)
 
 	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(int));
 	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
-	printf("Size of a char : %c byte(s)\n", (unsigned long)sizeof(char));
+	printf("Size of a char : %c byte(s)\n", (int)sizeof(char));
 	printf("Size of int r: %lu byte(s)\n", (unsigned long)sizeof(long int));
 	printf("Size of int j: %lu byte(s)\n", (unsigned long)sizeof(long long int));
 	return (0);
